+++ nest-forms-backend/src/admin/admin.controller.ts
@@ -10,7 +10,7 @@ import {
 
 import AdminGuard from '../auth/guards/admin.guard'
 import { ValidateFormRegistrationsResultDto } from '../nases/dtos/responses.dto'
-import NasesCronSubservice from '../nases/utils-services/nases.cron.subservice'
+import NasesCronService from '../nases/services/nases.cron.service'
 import { ErrorsEnum } from '../utils/global-enums/errors.enum'
 import ThrowerErrorGuard from '../utils/guards/thrower-error.guard'
 import AdminService from './admin.service'
@@ -22,7 +22,7 @@ export default class AdminController {
   constructor(
     private readonly adminService: AdminService,
     private readonly throwerErrorGuard: ThrowerErrorGuard,
-    private readonly nasesCronSubservice: NasesCronSubservice,
+    private readonly nasesCronService: NasesCronService,
   ) {}
 
   // Endpoints only for testing
@@ -87,6 +87,6 @@ export default class AdminController {
   @UseGuards(AdminGuard)
   @Get('check-form-registrations-in-nases')
   async checkFormsRegistrationsInNases(): Promise<ValidateFormRegistrationsResultDto> {
-    return this.nasesCronSubservice.validateFormRegistrations()
+    return this.nasesCronService.validateFormRegistrations()
   }
 }
